\name{DiVaMo-package}
\alias{DiVaMo-package}
\alias{DiVaMo}
\docType{package}
\title{
\packageTitle{DiVaMo}
}
\description{
\packageDescription{DiVaMo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DiVaMo}
\packageIndices{DiVaMo}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DiVaMo}

Maintainer: \packageMaintainer{DiVaMo}
}
\references{
~~ Literature or other references for background information ~~

~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
library("DiVaMo")
diVaMo()
}
